InfirienceProcessor/requirements.txt
matplotlib
requests
torch
numpy
opencv-python
torchvision
Pillow
Flask
flask-restful